chore: simplified readme page
